refactor: new keyboard hiding method and extract cell's resizing to the separate method
