+++ Agenda/Controllers/Agenda/AddGoalViewController.swift
@@ -44,6 +44,10 @@ final class AddGoalViewController: UIViewController {
         
         view.backgroundColor = #colorLiteral(red: 0.9490196078, green: 0.9490196078, blue: 0.968627451, alpha: 1)
         
+        // This methods is declared in Extensions/UIKit/UIViewController.swift
+        // It allows to hide keyboard when user taps in any place
+        hideKeyboardWhenTappedAround()
+        
         setupViewAndConstraints()
     }
     
@@ -143,7 +147,7 @@ private extension AddGoalViewController {
             tableView.contentInset = UIEdgeInsets(top: 0, left: 0, bottom: keyboardSize.height, right: 0)
         }
     }
-
+    
     @objc func keyboardWillHide(notification: NSNotification) {
         tableView.contentInset = .zero
     }
@@ -151,20 +155,7 @@ private extension AddGoalViewController {
 
 // MARK: - GoalTableViewCellDelegate
 extension AddGoalViewController: GoalTableViewCellDelegate {
-    // Update height of UITextView based on string height
     func updateHeightOfRow(_ cell: GoalTableViewCell, _ textView: UITextView) {
-        let size = textView.bounds.size
-        let newSize = tableView.sizeThatFits(CGSize(width: size.width, height: CGFloat.greatestFiniteMagnitude))
-        
-        if size.height != newSize.height {
-            UIView.setAnimationsEnabled(false)
-            tableView.beginUpdates()
-            tableView.endUpdates()
-            UIView.setAnimationsEnabled(true)
-            // Scoll up your textview if required
-            if let thisIndexPath = tableView.indexPath(for: cell) {
-                tableView.scrollToRow(at: thisIndexPath, at: .bottom, animated: false)
-            }
-        }
+        resize(cell, in: tableView, with: textView) // Update height of UITextView based on text's number of lines
     }
 }
