+++ Agenda/Extensions/UIKit/UIViewController.swift
@@ -1,0 +1,37 @@
+//
+//  UIViewController.swift
+//  Agenda
+//
+//  Created by Егор Бадмаев on 03.04.2022.
+//
+
+import UIKit
+
+extension UIViewController {
+    func hideKeyboardWhenTappedAround() {
+        let tap = UITapGestureRecognizer(target: self, action: #selector(UIViewController.dismissKeyboard))
+        tap.cancelsTouchesInView = false
+        
+        view.addGestureRecognizer(tap)
+    }
+    
+    @objc private func dismissKeyboard() {
+        view.endEditing(true)
+    }
+    
+    func resize(_ cell: GoalTableViewCell, in tableView: UITableView, with textView: UITextView) {
+        let size = textView.bounds.size
+        let newSize = tableView.sizeThatFits(CGSize(width: size.width, height: CGFloat.greatestFiniteMagnitude))
+        
+        if size.height != newSize.height {
+            UIView.setAnimationsEnabled(false)
+            tableView.beginUpdates()
+            tableView.endUpdates()
+            UIView.setAnimationsEnabled(true)
+            // Scoll up your textview if required
+            if let thisIndexPath = tableView.indexPath(for: cell) {
+                tableView.scrollToRow(at: thisIndexPath, at: .bottom, animated: false)
+            }
+        }
+    }
+}
